+++ sdk/src/utils/serde.rs
@@ -25,6 +25,35 @@ pub mod string {
     }
 }
 
+pub mod option_string {
+    use alloc::string::String;
+    use core::{fmt::Display, str::FromStr};
+
+    use serde::{de, Deserialize, Deserializer, Serializer};
+
+    pub fn serialize<T, S>(value: &Option<T>, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        T: Display,
+        S: Serializer,
+    {
+        match value {
+            Some(value) => serializer.collect_str(value),
+            None => serializer.serialize_none(),
+        }
+    }
+
+    pub fn deserialize<'de, T, D>(deserializer: D) -> Result<Option<T>, D::Error>
+    where
+        T: FromStr,
+        T::Err: Display,
+        D: Deserializer<'de>,
+    {
+        Option::<String>::deserialize(deserializer)?
+            .map(|string| string.parse().map_err(de::Error::custom))
+            .transpose()
+    }
+}
+
 pub mod option_prefix_hex_vec {
     use alloc::{string::String, vec::Vec};