+++ sdk/src/wallet/account/types/balance.rs
@@ -57,48 +57,6 @@ impl std::ops::AddAssign for Balance {
     }
 }
 
-/// Dto for the balance of an account, returned from [`crate::wallet::account::Account::sync()`] and
-/// [`crate::wallet::account::Account::balance()`].
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct BalanceDto {
-    /// Total and available amount of the base coin
-    pub base_coin: BaseCoinBalance,
-    /// Current required storage deposit amount
-    pub required_storage_deposit: RequiredStorageDeposit,
-    /// Native tokens
-    pub native_tokens: Vec<NativeTokensBalanceDto>,
-    /// Nfts
-    pub nfts: Vec<NftId>,
-    /// Aliases
-    pub aliases: Vec<AliasId>,
-    /// Foundries
-    pub foundries: Vec<FoundryId>,
-    /// Outputs with multiple unlock conditions and if they can currently be spent or not. If there is a
-    /// [`TimelockUnlockCondition`](crate::types::block::output::unlock_condition::TimelockUnlockCondition) or
-    /// [`ExpirationUnlockCondition`](crate::types::block::output::unlock_condition::ExpirationUnlockCondition) this
-    /// can change at any time
-    pub potentially_locked_outputs: HashMap<OutputId, bool>,
-}
-
-impl From<&Balance> for BalanceDto {
-    fn from(value: &Balance) -> Self {
-        Self {
-            base_coin: value.base_coin.clone(),
-            required_storage_deposit: value.required_storage_deposit.clone(),
-            native_tokens: value
-                .native_tokens
-                .iter()
-                .map(NativeTokensBalanceDto::from)
-                .collect::<_>(),
-            nfts: value.nfts.clone(),
-            aliases: value.aliases.clone(),
-            foundries: value.foundries.clone(),
-            potentially_locked_outputs: value.potentially_locked_outputs.clone(),
-        }
-    }
-}
-
 /// Base coin fields for [`Balance`]
 #[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize, CopyGetters)]
 #[serde(rename_all = "camelCase")]
@@ -156,59 +114,35 @@ pub struct NativeTokensBalance {
     /// Token id
     #[getset(get = "pub")]
     pub(crate) token_id: TokenId,
-    /// Token foundry immutable metadata
-    #[getset(get = "pub")]
-    pub(crate) metadata: Option<MetadataFeature>,
     /// Total amount
     #[getset(get_copy = "pub")]
     pub(crate) total: U256,
     /// Balance that can currently be spent
     #[getset(get_copy = "pub")]
     pub(crate) available: U256,
+    /// Token foundry immutable metadata
+    #[getset(get = "pub")]
+    #[serde(with = "crate::utils::serde::option_string")]
+    pub(crate) metadata: Option<MetadataFeature>,
 }
 
 impl Default for NativeTokensBalance {
     fn default() -> Self {
         Self {
             token_id: TokenId::null(),
-            metadata: None,
             total: U256::from(0u8),
             available: U256::from(0u8),
+            metadata: None,
         }
     }
 }
 
 impl std::ops::AddAssign for NativeTokensBalance {
     fn add_assign(&mut self, rhs: Self) {
-        if self.metadata.is_none() {
-            self.metadata = rhs.metadata;
-        }
         self.total += rhs.total;
         self.available += rhs.available;
-    }
-}
-
-/// Base coin fields for [`BalanceDto`]
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct NativeTokensBalanceDto {
-    /// Token id
-    pub token_id: TokenId,
-    /// Token foundry immutable metadata
-    pub metadata: Option<String>,
-    /// Total amount
-    pub total: U256,
-    /// Balance that can currently be spent
-    pub available: U256,
-}
-
-impl From<&NativeTokensBalance> for NativeTokensBalanceDto {
-    fn from(value: &NativeTokensBalance) -> Self {
-        Self {
-            token_id: value.token_id,
-            metadata: value.metadata.as_ref().map(|m| prefix_hex::encode(m.data())),
-            total: value.total,
-            available: value.available,
+        if self.metadata.is_none() {
+            self.metadata = rhs.metadata;
         }
     }
 }