+++ sdk/src/wallet/message_interface/message_handler.rs
@@ -38,7 +38,7 @@ use crate::{
             operations::transaction::{
                 high_level::minting::mint_native_token::MintTokenTransactionDto, TransactionOptions,
             },
-            types::{AccountIdentifier, BalanceDto, TransactionDto},
+            types::{AccountIdentifier, TransactionDto},
             OutputDataDto,
         },
         message_interface::{
@@ -519,6 +519,10 @@ impl WalletMessageHandler {
                 })
                 .await
             }
+            AccountMethod::ClaimableOutputs { outputs_to_claim } => {
+                let output_ids = account.claimable_outputs(outputs_to_claim).await?;
+                Ok(Response::OutputIds(output_ids))
+            }
             AccountMethod::ConsolidateOutputs {
                 force,
                 output_consolidation_threshold,
@@ -572,25 +576,19 @@ impl WalletMessageHandler {
                     .await?;
                 Ok(Response::GeneratedEvmAddresses(addresses))
             }
-            AccountMethod::SignEvm { message, chain } => {
+            AccountMethod::SignSecp256k1Ecdsa { message, chain } => {
                 let msg: Vec<u8> = prefix_hex::decode(message).map_err(crate::client::Error::from)?;
                 let (public_key, signature) = account
                     .wallet
                     .secret_manager
                     .read()
                     .await
-                    .sign_evm(&msg, &Chain::from_u32(chain))
+                    .sign_secp256k1_ecdsa(&msg, &Chain::from_u32(chain))
                     .await?;
-                Ok(Response::EvmSignature {
+                Ok(Response::Secp256k1EcdsaSignature {
                     public_key: prefix_hex::encode(public_key.to_bytes()),
                     signature: prefix_hex::encode(signature.to_bytes()),
                 })
-            }
-            AccountMethod::GetOutputsWithAdditionalUnlockConditions { outputs_to_claim } => {
-                let output_ids = account
-                    .get_unlockable_outputs_with_additional_unlock_conditions(outputs_to_claim)
-                    .await?;
-                Ok(Response::OutputIds(output_ids))
             }
             AccountMethod::GetOutput { output_id } => {
                 let output_data = account.get_output(&output_id).await;
@@ -729,7 +727,7 @@ impl WalletMessageHandler {
                 })
                 .await
             }
-            AccountMethod::GetBalance => Ok(Response::Balance(BalanceDto::from(&account.balance().await?))),
+            AccountMethod::GetBalance => Ok(Response::Balance(account.balance().await?)),
             AccountMethod::PrepareOutput {
                 params,
                 transaction_options,
@@ -783,9 +781,7 @@ impl WalletMessageHandler {
                 })
                 .await
             }
-            AccountMethod::SyncAccount { options } => {
-                Ok(Response::Balance(BalanceDto::from(&account.sync(options).await?)))
-            }
+            AccountMethod::SyncAccount { options } => Ok(Response::Balance(account.sync(options).await?)),
             AccountMethod::SendAmount { params, options } => {
                 convert_async_panics(|| async {
                     let transaction = account